+++ src/img/lib/cli.js
@@ -244,7 +244,7 @@ function CLI() {
  * By default a single line for an error is printed:
  *      imgadm: error (UnknownCommand): unknown command: "bogus"
  *
- * With one or more '-v' options a traceback is printed:
+ * With the '-v, --verbose' option a traceback is printed:
  *      imgadm: error (UnknownCommand): unknown command: "bogus"
  *
  *      UnknownCommandError: unknown command: "bogus"
@@ -315,10 +315,8 @@ CLI.prototype.main = function main(argv, options, callback) {
          * - If no `options.log` is given, we log to stderr.
          * - By default we log at the 'warn' level. Intentionally that is
          *   almost no logging
-         * - '-v|--verbose' to increase the logging level:
-         *      - 1: info
-         *      - 2: debug
-         *      - 3: trace and enable 'src' (source file location information)
+         * - '-v|--verbose' to set to trace and enable 'src' (source file
+         *   location information)
          * - '-E' to have a possible error be logged as the last single line
          *   of stderr as a Bunyan log record with an 'err'. I.e. in a
          *   structured format more useful to automation tooling.
@@ -327,7 +325,7 @@ CLI.prototype.main = function main(argv, options, callback) {
          * `bunyan` for readable output (at least until bunyan.js supports
          * doing it inline). Admittedly this is a bit of a pain:
          *
-         *      imgadm -vvv ... 2>&1 | bunyan
+         *      imgadm -v ... 2>&1 | bunyan
          */
         var log = options.log || bunyan.createLogger({
             name: self.name,
@@ -340,16 +338,8 @@ CLI.prototype.main = function main(argv, options, callback) {
             serializers: bunyan.stdSerializers
         });
         if (opts.verbose) {
-            var level;
-            if (opts.verbose.length === 1) {
-                level = 'info';
-            } else if (opts.verbose.length === 2) {
-                level = 'debug';
-            } else {
-                level = 'trace';
-                log = log.child({src: true});
-            }
-            log.level(level);
+            log = log.child({src: true});
+            log.level('trace');
         }
         self.log = log;
         self.verbose = Boolean(opts.verbose);
@@ -413,7 +403,7 @@ CLI.prototype.handleArgv = function handleArgv(argv, envopts, callback) {
     var longOpts = this.longOpts = {
         'help': Boolean,
         'version': Boolean,
-        'verbose': [Boolean, Array],
+        'verbose': Boolean,
         'E': Boolean
     };
     var shortOpts = this.shortOpts = {
@@ -453,7 +443,7 @@ CLI.prototype.printHelp = function printHelp(callback) {
         'Options:',
         '    -h, --help          Show this help message and exit.',
         '    --version           Show version and exit.',
-        '    -v, --verbose       Verbose logging. Multiple times for more.'
+        '    -v, --verbose       Verbose logging.'
     ]);
 
     if (self.envopts && self.envopts.length) {
@@ -1321,7 +1311,7 @@ CLI.prototype.do_update.description = (
 
 
 /**
- * `imgadm create [<options>] <uuid> [<manifest-field>=<value> ...]`
+ * `imgadm create [<options>] <vm-uuid> [<manifest-field>=<value> ...]`
  */
 CLI.prototype.do_create = function do_create(subcmd, opts, args, callback) {
     var self = this;
@@ -1473,8 +1463,7 @@ CLI.prototype.do_create = function do_create(subcmd, opts, args, callback) {
 };
 CLI.prototype.do_create.description = (
     /* BEGIN JSSTYLED */
-    '**Experimental. This command currently does not work on KVM zones.**\n'
-    + 'Create a new image from a prepared and stopped VM.\n'
+    'Create a new image from a prepared and stopped VM.\n'
     + '\n'
     + 'To create a new virtual image, one first creates a VM from an existing\n'
     + 'image, customizes it, runs "sm-prepare-image", shuts it down, and\n'
@@ -1504,7 +1493,7 @@ CLI.prototype.do_create.description = (
     + '    -c COMPRESSION One of "none", "gz" or "bzip2" for the compression\n'
     + '                   to use on the image file, if any. Default is "none".\n'
     + '    -i             Build an incremental image (based on the "@final"\n'
-    + '                   snapshot of the source image for VM.\n'
+    + '                   snapshot of the source image for the VM).\n'
     + '\n'
     + '    -p URL, --publish URL\n'
     + '                   Publish directly to the given image source\n'